tetragon/windows: Port Tetragon on Windows
This PR renames main_linux.go to main.go to use common code between Windows and Linux

btf, profs, namespaces, struct aligmnment based packages are not available in Windows.
Helper functions using these packages are implemented in main_linux.go and stubbed out in main_windows.go
Logically, there is no difference in main.go between Linux and Windows.
