+++ cmd/tetragon/main_windows.go
@@ -1,0 +1,26 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Tetragon
+
+package main
+
+func logCurrentSecurityContext() {
+}
+
+func initHostNamespaces() error {
+	return nil
+}
+
+func checkProcFS() {
+
+}
+
+func initCachedBTF(lib, btf string) error {
+	return nil
+}
+
+func checkStructAlignments() error {
+	return nil
+}
+
+func setNetNSDir() {
+}
